Cache hit realization

+++ ht-functions.c
@@ -1,5 +1,5 @@
 //simple integer hash function with prime module
-int hash(int data)
+long hash(long data)
 {
     return data % modular;
 }
